[5:10:00] to [5:20:00]
Transcript between the times [5:10:00] and [5:20:00].
It's untested, at all.
And I have a doubt : the voice said "setheading O" and "forward O" (both times like the letter O), but it looked like it should be zero instead. So that's what I wrote. If it's not the case, be sure to correct it.