Remove interpreter-specific info, explain NOTEs
Now that there are a good many logos that are known to work with
the unchanged text, I can remove a few helper lines that were
intended for various logo implementaitons.

Also added some annotations for where changes from the original
text have been made, and a line at the top to point out that
NOTEs are not in the original text. This is to address #57.